Rename Forms.tsx to Fields.tsx and add documentation
